requirements.txt
click==8.0.4
Flask==2.0.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
psycopg2-binary==2.9.3
Werkzeug==2.0.3
